requirements.txt
# Food Trends Tracker - SERP API + Groq

groq>=0.4.0
google-search-results>=2.4.2
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0

